refactor(generate): 退役双源旧模块——扫盘索引 / cache index / save 端点
出图时间线切 DB 单源(前一 commit)后,旧「disk 扫盘 ∪ cache index」双源
链路已无消费者,整体退役:

- 删 services/generate_history_index.py(#475 SQLite 扫盘索引,465 行):
  列表来源已是 tasks 台账,PNG chunk reader / v1→v2 迁移随之无消费方
- 删 POST /api/generate/save(multipart 上传落盘,~200 行):写路径已全部
  在 server 端闭环,前端零上传;xy tmp 文件夹清理保留为老版本残留兼容
- 删 GET /api/generate/disk/history、GET /api/generate/cache/index、
  DELETE /api/generate/disk/*(删除统一走 DELETE /api/queue/{id})
- disk_cache 删 list_index 聚合与 mode/xy_info 字段:cache 降级为纯图
  字节存储,列表/参数权威在 tasks 台账
- client.ts 删 listDiskGenerateHistory/listCacheGenerateHistory 及类型
- 测试:test_generate_save_metadata.py 重写为 test_generate_png_metadata.py
  (写路径改走 generate_storage 直落,保留 a1111/Civitai hash 链路与
  disk image/thumb 端点覆盖);test_generate_history_index.py 删除
