+++ studio/web/src/api/client.ts
@@ -1572,47 +1572,6 @@ export interface GenerateTimelineEntry {
   composite_url?: string
 }
 
-/** GET /api/generate/cache/index — 当前 session 加密磁盘 cache 索引。
- *  server 端 SessionCache 按 task_id 聚合返回；前端转成 CacheEntry。 */
-export interface CacheGenerateHistoryEntry {
-  /** "cache:<task_id>" */
-  id: string
-  taskId: number
-  mode: 'single' | 'xy'
-  /** Unix timestamp ms */
-  createdAt: number
-  /** 该 task 的所有文件名（XY 时按文件名排序） */
-  filenames: string[]
-  /** GenerateParamsSnapshot dict */
-  params: Record<string, unknown>
-  /** 仅 mode=xy 存在；列每张图的 xy 位置，PreviewXYGrid 重建网格用 */
-  samples?: Array<{
-    filename: string
-    xy: { xi: number; yi: number; xv: string | number; yv: string | number | null }
-  }>
-}
-
-/** 落盘测试图历史 entry（GET /api/generate/disk-history）。
- *  params 是 GenerateParamsSnapshot（前端用 paramsSnapshot.ts 的类型解读），
- *  这里用 unknown 让 api/client.ts 不依赖 pages 层类型。 */
-export interface DiskGenerateHistoryEntry {
-  /** 稳定 ID："disk:<date>:<mode>:image_<N>"；前端按此 dedup */
-  id: string
-  /** YYYY-MM-DD */
-  date: string
-  mode: 'single' | 'xy'
-  filename: string
-  /** 服务端绝对路径，用于和 IDB entry.diskPath 做 dedup */
-  path: string
-  /** /api/generate/disk-image/<date>/<mode>/<filename> */
-  url: string
-  /** Unix timestamp（sidecar 写入或 fallback 文件 mtime） */
-  created_at: number
-  schema_version: number
-  /** sidecar 里的 params object（前端按 GenerateParamsSnapshot 解读） */
-  params: Record<string, unknown>
-}
-
 /** version output/ 下扫到的 training_state_step*.pt（断点续训用）。 */
 export interface StateCkpt {
   /** global_step 数 */
@@ -2791,15 +2750,6 @@ export const api = {
   /** PR-9 — 启动测试出图 task。Phase 2 起：图走 server 内存 cache，关页面即丢。 */
   enqueueGenerate: (body: GenerateRequest) =>
     req<Task>('/api/generate', { method: 'POST', body: JSON.stringify(body) }),
-  /** 落盘历史：扫 studio_data/test/&lt;date&gt;/{single,xy}/image_N.json sidecar，
-   *  按 created_at desc 返回；用于历史栏跨会话回看已落盘的测试图。
-   *  注意路径用 disk/ 子前缀避开 `/api/generate/{task_id}` 的单段 catch-all。 */
-  listDiskGenerateHistory: (limit = 500) =>
-    req<{ entries: DiskGenerateHistoryEntry[] }>(`/api/generate/disk/history?limit=${limit}`),
-  /** 当前 session 加密磁盘 cache 历史（save_test_images=false 时唯一来源）。
-   *  server 重启 / SSE 断连 30s + LRU 后 entry 消失；刷新 / 切路由都拉这里。 */
-  listCacheGenerateHistory: () =>
-    req<{ entries: CacheGenerateHistoryEntry[] }>('/api/generate/cache/index'),
   /** 出图时间线（DB 单源）：所有 generate 任务行，id desc 分页。 */
   listGenerateTimeline: (limit = 500, offset = 0) =>
     req<{ entries: GenerateTimelineEntry[]; total: number; offset: number }>(
